+++ lib/Fluence.Wpf/Fluence.Wpf/Automation/AutoSuggestBoxAutomationPeer.cs
@@ -27,6 +27,7 @@
  */
 
 using Fluence.Wpf.Controls;
+using System.Windows.Automation;
 using System.Windows.Automation.Peers;
 using System.Windows.Automation.Provider;
 
@@ -72,12 +73,22 @@ namespace Fluence.Wpf.Automation
         /// <inheritdoc />
         public virtual string Value => AutoSuggestBox.Text;
 
-        /// <inheritdoc />
-        public virtual bool IsReadOnly => !AutoSuggestBox.IsEnabled;
+        /// <summary>
+        /// Always <see langword="false"/>. <see cref="AutoSuggestBox"/> has no read-only mode;
+        /// disabled state is conveyed via <see cref="System.Windows.UIElement.IsEnabled"/>,
+        /// not <see cref="IValueProvider.IsReadOnly"/>.
+        /// </summary>
+        public virtual bool IsReadOnly => false;
 
         /// <inheritdoc />
+        /// <exception cref="ElementNotEnabledException">The control is disabled.</exception>
         public virtual void SetValue(string value)
         {
+            if (!IsEnabled())
+            {
+                throw new ElementNotEnabledException();
+            }
+
             AutoSuggestBox.Text = value;
         }