+++ lib/Fluence.Wpf/Fluence.Wpf/Automation/DropDownButtonAutomationPeer.cs
@@ -49,7 +49,7 @@ namespace Fluence.Wpf.Automation
         /// <inheritdoc />
         protected override AutomationControlType GetAutomationControlTypeCore()
         {
-            return AutomationControlType.SplitButton;
+            return AutomationControlType.Button;
         }
 
         /// <inheritdoc />